fixed solution to task 14

+++ task14.py
@@ -1,3 +1,3 @@
-x = int(input())
+x = float(input())
 #print(((x+2)*3-6)/3-4)
 print(x-4)
